Set the LOGNAME from whoami
Well, $LOGNAME is defined by POSIX (see
a801c4c58902cae7b835d890487edc39bd7da142) but it's not appearing in
reality under clean environments:

    ~% env -i /bin/bash
    mike:/home/mike$ echo $LOGNAME

    mike:/home/mike$ exit
    ~% env -i /bin/dash
    $ echo $LOGNAME

    $ exit
    ~% env -i /bin/ksh
    $ echo $LOGNAME

    $ exit
    ~% env -i /usr/bin/zsh
    ~% echo $LOGNAME
    mike
    ~% exit

Fall back to whoami(1) when $LOGNAME is not set.

Modify one of the tests to run with a restricted environment in order to
test this. I had tried modifying the test-driver to run all tests in a
restricted environment but the test-driver script is generated by
Automake.

Closes #165.